Trial: parameterize test by individual tree nodes
paramaterizing by node seems like it would be nice for reporting,
but doing this for single test took the suite from
~2 seconds to 15+ minutes, using 10+ GB of memory.
I'm thinking the test suite will have to err on the side of
"finding problems quickly" rather than "exhaustive error reporting".
The latter would need to be reserved for manual queries when trying to
solve a problem.